requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dlib==19.24.2
dnspython==2.4.2
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==3.0.0
Flask-Cors==4.0.0
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.24.4
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.1.0
pymongo==4.6.1
pytesseract==0.3.10
Werkzeug==3.0.1
zipp==3.17.0
